# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "STD Plus API",
    "version": "1.0.0",
    "description": "Backend API for STD Plus — materials, inventory, orders, and workflow management."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string" },
          "data": {}
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string" }
        }
      },
...
      "DeleteManyBody": {
        "type": "object",
        "required": ["ids"],
        "properties": {
          "ids": {
            "type": "array",
            "items": { "type": "string" },
            "minItems": 1
          }
        }
      },
      "DeleteManyResult": {
        "type": "object",
        "properties": {
          "deletedCount": { "type": "integer" }
        }
      },
...
      "Dimensions": {
        "type": "object",
        "properties": {
          "width": { "type": "number", "minimum": 0 },
          "height": { "type": "number", "minimum": 0 },
          "thickness": { "type": "number", "minimum": 0 }
        }
      },
...
      "PaneCreateBody": {
        "type": "object",
        "required": ["request"],
        "properties": {
          "request": { "type": "string", "minLength": 1, "description": "Request ID this pane belongs to" },
          "order": { "type": "string", "minLength": 1, "description": "Order ID (optional, auto-filled when order is created from request)" },
          "material": { "type": "string", "minLength": 1, "description": "Material ID (optional, auto-filled from the order's material when order is created from request)" },
          "inventory": { "type": "string", "minLength": 1, "description": "Inventory slot ID. If provided, a 'cut' MaterialLog is auto-created for traceability." },
          "currentStation": { "type": "string", "minLength": 1, "description": "Station ObjectId" },
          "currentStatus": { "type": "string", "enum": ["pending", "in_progress", "awaiting_scan_out", "completed", "claimed", "defected"] },
          "routing": { "type": "array", "items": { "type": "string", "minLength": 1 }, "description": "Array of Station ObjectIds" },
          "customRouting": { "type": "boolean" },
          "dimensions": { "$ref": "#/components/schemas/Dimensions" },
          "jobType": { "type": "string", "description": "What the customer wants done (e.g. Laminated, Tempered, Beveled)" },
          "rawGlass": { "type": "object", "properties": { "glassType": { "type": "string" }, "color": { "type": "string" }, "thickness": { "type": "number", "minimum": 0 }, "sheetsPerPane": { "type": "integer", "minimum": 1 } } },
          "glassType": { "type": "string" },
          "glassTypeLabel": { "type": "string" },
          "cornerSpec": { "type": "string", "description": "Corner cutting specification" },
          "dimensionTolerance": { "type": "string", "description": "Acceptable dimension tolerance" },
          "holes": { "type": "array", "items": { "$ref": "#/components/schemas/HoleNotch" }, "description": "Array of hole positions and dimensions" },
          "notches": { "type": "array", "items": { "$ref": "#/components/schemas/HoleNotch" }, "description": "Array of notch positions and dimensions" },
          "processes": { "type": "array", "items": { "type": "string" }, "description": "Processing steps required (e.g. temper, laminate, assembly, cnc)" },
          "edgeTasks": { "type": "array", "items": { "$ref": "#/components/schemas/EdgeTask" } },
          "withdrawal": { "type": "string", "minLength": 1 },
          "remakeOf": { "type": "string", "minLength": 1 },
          "laminateRole": { "type": "string", "enum": ["single", "parent", "sheet"] },
          "parentPane": { "type": "string", "minLength": 1 },
          "childPanes": { "type": "array", "items": { "type": "string", "minLength": 1 } },
          "sheetLabel": { "type": "string" },
          "laminateStation": { "type": "string", "minLength": 1 },
          "startedAt": { "type": "string", "format": "date-time" },
          "completedAt": { "type": "string", "format": "date-time" },
          "deliveredAt": { "type": "string", "format": "date-time" },
          "notes": { "type": "string" }
        }
      },
...
    },
    "parameters": {
      "IdParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "Resource ID"
      },
...
    },
    "responses": {
      "401": {
        "description": "Not authenticated",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "404": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "400": {
        "description": "Validation failed",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
          }
        }
      },
      "403": {
        "description": "Not authorized for this action",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "409": {
        "description": "Cannot delete — referenced by other records",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/panes": {
      "get": {
        "tags": ["Panes"],
        "summary": "List all panes",
        "description": "Returns panes matching optional filters. Does **not** use cursor/offset pagination — returns up to `limit` results (default 200, max 1000) sorted by `sort` (default `-createdAt`). By default excludes `currentStatus: merged_into` (use `includeMerged=true` to include retired laminate rows).",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "order", "in": "query", "schema": { "type": "string" }, "description": "Filter by order ID" },
          { "name": "request", "in": "query", "schema": { "type": "string" }, "description": "Filter by request ID" },
          { "name": "material", "in": "query", "schema": { "type": "string" }, "description": "Filter by material ID" },
          { "name": "station", "in": "query", "schema": { "type": "string" }, "description": "Filter by current station ObjectId. Use the string `null` to find panes in virtual states (queued, completed, claimed)." },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["pending", "in_progress", "awaiting_scan_out", "completed", "claimed"] }, "description": "Filter by exact current status" },
          { "name": "status_ne", "in": "query", "schema": { "type": "string", "enum": ["pending", "in_progress", "awaiting_scan_out", "completed", "claimed"] }, "description": "Exclude panes with this status (not-equal filter). Useful for excluding claimed panes from order views: `?order=<id>&status_ne=claimed`. If both `status` and `status_ne` are provided, `status_ne` takes precedence." },
          { "name": "laminateRole", "in": "query", "schema": { "type": "string", "enum": ["single", "parent", "sheet"] }, "description": "Filter by laminate role. Use `parent` to get only parent panes, `sheet` for raw sheets, `single` for non-laminate panes." },
          { "name": "parentPane", "in": "query", "schema": { "type": "string" }, "description": "Filter by parent pane ID. Returns all child sheets belonging to a specific parent pane." },
          { "name": "includeMerged", "in": "query", "schema": { "type": "string", "enum": ["true", "false"] }, "description": "When `true`, include `merged_into` panes. Default omits them unless `status` or `status_ne` is set." },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 200, "maximum": 1000, "minimum": 1 }, "description": "Max results to return (default 200, max 1000)" },
          { "name": "sort", "in": "query", "schema": { "type": "string", "default": "-createdAt" }, "description": "Sort field (prefix with `-` for descending)" }
        ],
        "responses": {
          "200": { "description": "List of panes", "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/SuccessResponse" }, { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Pane" } } } }] } } } },
          "401": { "$ref": "#/components/responses/401" }
        }
      },
      "post": {
        "tags": ["Panes"],
        "summary": "Create a pane",
        "description": "Requires `panes:create` permission.\n\n**Laminate auto-splitting:** When `rawGlass.sheetsPerPane` is greater than 1 **and** the `routing` array includes a station with `isLaminateStation: true`, the endpoint automatically creates a laminate pane structure instead of a single pane:\n- 1 **parent** pane (`laminateRole: \"parent\"`) with routing set to post-lamination stations only, `currentStation: null`, `currentStatus: \"pending\"`\n- N **sheet** panes (`laminateRole: \"sheet\"`) with routing set to pre-lamination stations (up to and including the laminate station), each starting at the first station\n- Parent and sheets are linked via `childPanes` / `parentPane`\n\nThe response shape changes to `{ parent: Pane, sheets: Pane[] }` instead of a single `Pane`.\n\nReturns `400` if `sheetsPerPane > 1` but no laminate station is found in the routing.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PaneCreateBody" } } } },
        "responses": {
          "201": {
            "description": "Pane created. For regular panes, `data` is a single Pane object. For laminate panes (`sheetsPerPane > 1`), `data` is `{ parent: Pane, sheets: Pane[] }`.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "oneOf": [
                            { "$ref": "#/components/schemas/Pane", "description": "Single pane (regular)" },
                            {
                              "type": "object",
                              "description": "Laminate pane structure (when sheetsPerPane > 1)",
                              "properties": {
                                "parent": { "$ref": "#/components/schemas/Pane" },
                                "sheets": { "type": "array", "items": { "$ref": "#/components/schemas/Pane" } }
                              },
                              "required": ["parent", "sheets"]
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/400" },
          "401": { "$ref": "#/components/responses/401" },
          "403": { "$ref": "#/components/responses/403" }
        }
      },
      "delete": {
        "tags": ["Panes"],
        "summary": "Delete multiple panes",
        "description": "Requires `panes:manage` permission.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeleteManyBody" } } } },
        "responses": {
          "200": { "description": "Panes deleted", "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/SuccessResponse" }, { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/DeleteManyResult" } } }] } } } },
          "401": { "$ref": "#/components/responses/401" },
          "403": { "$ref": "#/components/responses/403" },
          "409": { "$ref": "#/components/responses/409" }
        }
      }
    },
    "/api/panes/{id}": {
      "get": {
        "tags": ["Panes"],
        "summary": "Get pane by ID or pane number",
        "description": "Accepts either a MongoDB ObjectId or a pane number (e.g. `PNE-0001`). The endpoint auto-detects the format and queries accordingly. **Pane numbers** are matched **case-insensitively** (e.g. `pne-0001` and `PNE-0001` resolve to the same document).",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Pane ObjectId or pane number (e.g. `PNE-0001`)", "example": "PNE-0001" }],
        "responses": {
          "200": { "description": "Pane found", "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/SuccessResponse" }, { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Pane" } } }] } } } },
          "401": { "$ref": "#/components/responses/401" },
          "404": { "$ref": "#/components/responses/404" }
        }
      },
      "patch": {
        "tags": ["Panes"],
        "summary": "Update a pane",
        "description": "Requires `panes:manage` permission.\n\n**Laminate auto-splitting:** When a pane with `laminateRole: \"single\"` is updated with `routing` that includes a laminate station, and `rawGlass.sheetsPerPane` is greater than 1 (either from the update body or from the existing pane data), the endpoint automatically splits the pane into a laminate structure:\n- The existing pane becomes the **parent** (`laminateRole: \"parent\"`) with routing set to post-lamination stations only\n- N **sheet** panes are created (`laminateRole: \"sheet\"`) with routing up to and including the laminate station, each starting at the first station\n- Parent and sheets are linked via `childPanes` / `parentPane`\n- Response shape becomes `{ parent: Pane, sheets: Pane[] }`\n\nThis is the typical flow: panes are created without routing at Order Release, then when routing is assigned during release, the laminate split happens automatically.\n\n**Order release side effect:** When a pane's `order` field is set and the pane was previously unlinked (`order: null`), the target order's `paneCount` is automatically incremented and `progressPercent` is recalculated. This is the mechanism used to release a remake pane into an order — the remake pane is created with `order: null` during claim approval, then later assigned to the order via this endpoint, which keeps the order's pane count accurate. A real-time `order:updated` event is emitted.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdParam" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PaneUpdateBody" } } } },
        "responses": {
          "200": {
            "description": "Pane updated. For regular updates, `data` is a single Pane object. When laminate split is triggered, `data` is `{ parent: Pane, sheets: Pane[] }`.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "oneOf": [
                            { "$ref": "#/components/schemas/Pane", "description": "Single pane (regular update)" },
                            {
                              "type": "object",
                              "description": "Laminate pane structure (when split is triggered)",
                              "properties": {
                                "parent": { "$ref": "#/components/schemas/Pane" },
                                "sheets": { "type": "array", "items": { "$ref": "#/components/schemas/Pane" } }
                              },
                              "required": ["parent", "sheets"]
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/401" },
          "403": { "$ref": "#/components/responses/403" },
          "404": { "$ref": "#/components/responses/404" }
        }
      },
      "delete": {
        "tags": ["Panes"],
        "summary": "Delete a pane",
        "description": "Requires `panes:manage` permission.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdParam" }],
        "responses": {
          "200": { "description": "Pane deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "401": { "$ref": "#/components/responses/401" },
          "403": { "$ref": "#/components/responses/403" },
          "404": { "$ref": "#/components/responses/404" },
          "409": { "$ref": "#/components/responses/409" }
        }
      }
    },
...
    "/api/panes/{paneNumber}/scan": {
      "post": {
        "tags": ["QR Scanning"],
        "summary": "Scan a pane (QR scan endpoint)",
        "description": "Core endpoint for the QR scanning workflow. Accepts a pane number (from the QR code) and an action to perform at a given station.\n\n**QR codes** contain values like `STDPLUS:PNE-0001`. The frontend should strip the `STDPLUS:` prefix and send only `PNE-0001` as the `paneNumber` path parameter.\n\n**Actions:**\n- `scan_in` — Sets `currentStation` to the station, `currentStatus` to `in_progress`, and `startedAt` (if not already set). For sheet panes at a laminate station, emits `laminate:ready` or `laminate:waiting` WebSocket events.\n- `start` — Sets `currentStation` to the station and `currentStatus` to `in_progress`.\n- `complete` — Sets `currentStation` to the station and `currentStatus` to `awaiting_scan_out`.\n- `scan_out` — Requires status `awaiting_scan_out`. Advances the pane to the next station in its routing with status `pending`. If it's the last station, pane status becomes `completed` and `completedAt` is set. Sheet panes do not update `panesCompleted`/`progressPercent`.\n- `qc_pass` — Same rules as `scan_out`. PaneLog records `qc_pass` for QC audit; routing and order progress match `scan_out`.\n- `qc_fail` — Requires `awaiting_scan_out`. Marks the pane `defected`, logs `qc_fail` with required `reason` and optional `description`, and auto-creates a remake pane (returned as `remadePane`). Optional `remakeStationId` queues the remake at that station; otherwise the first routing station is used.\n- `laminate` — Merges sheets at lamination. **Recommended:** call with a **sheet** pane number in the path; optional `laminateSurvivorPaneNumber` picks which sheet keeps its QR (defaults to path pane). If the path is the **dormant parent**, `laminateSurvivorPaneNumber` is **required**. All active sheets (`!= claimed`, `!= merged_into`) must be at the lamination station in `in_progress` or `awaiting_scan_out`. **Survivor** becomes `laminateRole: single` with post-lam `routing`, `awaiting_scan_out` at lamination; other sheets and the parent row become `merged_into` → survivor. Worker then `scan_out` / `qc_pass` using the **survivor** QR. First `scan_out` from lam jumps to `routing[0]` (lam station is not in survivor routing). Returns `mergedSheets`, `survivorPaneNumber`, `retiredPaneNumbers`, `parentRetired`. Scanning a `merged_into` pane returns **400** with `errors.code: MERGED_INTO`.\n\n**Side effects (every action):**\n- Creates a `PaneLog` entry.\n- Emits `pane:updated` with `action: \"scanned\"` to `dashboard`, `pane`, `production`, `station:<name>` (on `qc_fail`, also `qc_failed` for the defected pane and `created` for the remake).\n- Emits `log:updated` with `action: \"pane_scanned\"` to `log`.\n- If the pane has no `material` but has an `order`, the order's material is backfilled onto the pane.\n\n**Additional side effects (`scan_out` / `qc_pass`):**\n- Updates associated `Order` progress (`panesCompleted`, `progressPercent`, `stationBreakdown`) and order `status`.\n- Emits `order:updated`, `station:pane_arrived`, and `notification` WebSocket events.\n- Creates a `Notification` for the order's assigned worker (if any).\n- Emits a station-room `notification` event to `station:<nextStation>` for the pane arrival.\n\n**Additional side effects (`qc_fail`):**\n- Decrements `stationBreakdown` at the QC station; may bump `order.paneCount` for non-sheet panes.\n- May create a `MaterialLog` with `referenceType: \"qc_remake\"`.\n- Emits `order:updated`, remake queue `station:pane_arrived` / station notifications, and user `notification` (`qc_remake`) when applicable.\n\n**Validations:**\n- `complete`, `scan_out`, `qc_pass`, and `qc_fail` must be performed at the pane's `currentStation` when that applies.\n- `scan_out` and `qc_pass` require pane status `awaiting_scan_out` (i.e. `complete` must be called first).\n- `qc_fail` requires `reason`; pane must be eligible to fail (e.g. not already `defected` / `completed` / `claimed`).\n- Scanning a pane that is already `completed` (for non-QC-fail paths) returns 400.\n- `laminate` requires a sheet or parent pane in the path; all active child sheets must be at the lamination station; parent path requires `laminateSurvivorPaneNumber`.\n\nRequires `panes:scan` permission.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "paneNumber",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "The pane number (e.g. `PNE-0001`). Strip the `STDPLUS:` prefix from QR codes before sending.",
            "example": "PNE-0001"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ScanBody" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scan processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": { "$ref": "#/components/schemas/ScanResult" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "description": "Validation error — wrong station, already completed, laminate rules, or scan on `merged_into` pane (`errors.code`: `MERGED_INTO`, `survivorPaneNumber`)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "$ref": "#/components/responses/401" },
          "404": { "description": "Pane not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/panes/batch-scan": {
      "post": {
        "tags": ["QR Scanning"],
        "summary": "Batch scan multiple panes",
        "description": "Performs the same scan action on multiple panes in one request. Useful for scanning a group of panes at the same station simultaneously.\n\n**Supported actions:** `scan_in`, `start`, `complete`. **`scan_out`**, **`qc_pass`**, **`qc_fail`**, and **`laminate`** are **not** supported in batch mode — use the single-pane scan endpoint for those.\n\n**Behavior:**\n- **Application rollback (no MongoDB transaction):** If any pane fails mid-batch (e.g. wrong station for `complete`, already `completed`, **`merged_into`**), earlier writes in that request are reverted (pane fields restored from a snapshot, `PaneLog` rows created in the batch removed). Works on standalone MongoDB.\n- Panes are processed in the order they appear in `paneNumbers` (missing numbers are skipped; if none match, **400**).\n- For `scan_in`/`start`: Sets `currentStation` and `currentStatus` to `in_progress`. Sets `startedAt` if not already set.\n- For `complete`: Validates the pane is at the specified station. Sets `currentStatus` to `awaiting_scan_out`.\n- Creates a `PaneLog` entry for each pane.\n- If a pane has no `material` but has an `order`, the order's material is backfilled.\n- On success, emits `pane:updated` and `log:updated` WebSocket events for each pane.\n\nRequires `panes:scan` permission.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["paneNumbers", "station", "action"],
                "properties": {
                  "paneNumbers": {
                    "type": "array",
                    "items": { "type": "string", "minLength": 1 },
                    "minItems": 1,
                    "description": "Array of pane numbers to scan (e.g. [\"PNE-0001\", \"PNE-0002\"])"
                  },
                  "station": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Station ObjectId where the scan is performed"
                  },
                  "action": {
                    "type": "string",
                    "enum": ["scan_in", "start", "complete"],
                    "description": "Action to perform on all panes"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch scan processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "updatedCount": { "type": "integer", "description": "Number of panes updated" },
                            "panes": { "type": "array", "items": { "$ref": "#/components/schemas/Pane" }, "description": "Array of updated pane objects (populated)" }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "description": "Validation error — unsupported action, empty pane list, no matching pane numbers, merged/completed/wrong-station pane in batch, or invalid body", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "$ref": "#/components/responses/401" },
          "500": { "description": "Unexpected server error (e.g. failure while applying rollback)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
  },
...
}